refactor: separate git utils into a separate package
